fixd a lot of warnings, we are now almost < 100 for the whole project.
git-svn-id: https://rubyeclipse.svn.sourceforge.net/svnroot/rubyeclipse/trunk@2054 0bd42bcd-5f16-0410-9896-c9ec0b0ce843

+++ org.rubypeople.rdt.refactoring/src/org/rubypeople/rdt/refactoring/util/NameHelper.java
@@ -48,7 +48,7 @@ public class NameHelper {
 	public static String createName(String string) {
 		Matcher matcher = Pattern.compile("([@]{0,2}\\w*[a-zA-Z_])(\\d+)").matcher(string);
 		if (matcher.matches()) {
-			return matcher.group(1) + String.valueOf(Integer.valueOf(matcher.group(2)) + 1);
+			return matcher.group(1) + String.valueOf(Integer.valueOf(matcher.group(2)).intValue() + 1);
 		}
 		return string + 1;
 	}
